requirements.txt
openai
opencv-python
numpy
Pillow
torch
